create shared code file and update docs
